+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file, following t
 
 ## [Unreleased]
 
+- Zoom event gives more info and is customizable by `setAlignment`
+
 ## [0.3.0] - 2024-02-21
 
 - Customizing tooltips via `setTooltip`